+++ addon/query/indexeddb-adapter.js
@@ -6,8 +6,9 @@ import Ember from 'ember';
 import FilterOperator from './filter-operator';
 import { SimplePredicate, ComplexPredicate, StringPredicate, DetailPredicate } from './predicate';
 import BaseAdapter from './base-adapter';
-import { getAttributeFilterFunction, buildProjection, buildOrder, buildTopSkip } from './js-adapter';
+import { getAttributeFilterFunction, buildProjection, buildOrder, buildTopSkip, buildFilter } from './js-adapter';
 import Information from '../utils/information';
+import Dexie from 'npm:dexie';
 
 /**
   Class of query language adapter that allows to load data from IndexedDB.
@@ -41,34 +42,52 @@ export default class extends BaseAdapter {
   */
   query(query) {
     return new Ember.RSVP.Promise((resolve, reject) => {
+      let isBadQuery = containsRelationships(query);
       let order = buildOrder(query);
       let topskip = buildTopSkip(query);
-      let projection = buildProjection(query);
+      let jsProjection = buildProjection(query);
       let table = this._db.table(query.modelName);
+      let filter = query.predicate ? buildFilter(query.predicate, { booleanAsString: true }) : (data) => data;
+      let projection = query.projectionName ? query.projectionName : query.projection ? query.projection : null;
 
-      updateWhereClause(table, query).toArray().then((data) => {
-        let response = { meta: {}, data: projection(topskip(order(data))) };
-        if (query.count) {
-          response.meta.count = data.length;
+      (isBadQuery ? table : updateWhereClause(table, query)).toArray().then((data) => {
+        let length = data.length;
+        if (!isBadQuery) {
+          data = topskip(order(data));
         }
 
-        resolve(response);
-      }).catch((error) => {
-        reject(error);
-      });
+        Dexie.Promise.all(data.map(i => i.loadRelationships(projection))).then(() => {
+          if (isBadQuery) {
+            data = filter(data);
+            length = data.length;
+            data = topskip(order(data));
+          }
+
+          if (!projection) {
+            data = jsProjection(data);
+          }
+
+          let response = { meta: {}, data: data };
+          if (query.count) {
+            response.meta.count = length;
+          }
+
+          resolve(response);
+        }, reject);
+      }, reject);
     });
   }
 }
 
 /**
- * Builds Dexie `WhereClause` for filtering data.
- * Filtering only with Dexie can applied only for simple cases (for `SimplePredicate`).
- * For complex cases all logic implemened programmatically.
- *
- * @param {Dexie.Table} table Table instance for loading objects.
- * @param {Query} query Query language instance for loading data.
- * @returns {Dexie.Collection|Dexie.Table} Table or collection that can be used with `toArray` method.
- */
+  Builds Dexie `WhereClause` for filtering data.
+  Filtering only with Dexie can applied only for simple cases (for `SimplePredicate`).
+  For complex cases all logic implemened programmatically.
+
+  @param {Dexie.Table} table Table instance for loading objects.
+  @param {Query} query Query language instance for loading data.
+  @returns {Dexie.Collection|Dexie.Table} Table or collection that can be used with `toArray` method.
+*/
 function updateWhereClause(table, query) {
   let predicate = query.predicate;
 
@@ -85,9 +104,8 @@ function updateWhereClause(table, query) {
   }
 
   if (predicate instanceof SimplePredicate) {
-    let fields = Information.parseAttributePath(predicate.attributePath);
     let value = typeof predicate.value === 'boolean' ? `${predicate.value}` : predicate.value;
-    if (value === null || fields.length > 1) {
+    if (value === null) {
       // IndexedDB (and Dexie) doesn't support null - use JS filter instead.
       // https://github.com/dfahlander/Dexie.js/issues/153
       return table.filter(getAttributeFilterFunction(predicate));
@@ -117,9 +135,54 @@ function updateWhereClause(table, query) {
     }
   }
 
-  if (predicate instanceof StringPredicate || predicate instanceof DetailPredicate || predicate instanceof ComplexPredicate) {
-    return table.filter(getAttributeFilterFunction(predicate));
+  if (predicate instanceof StringPredicate || predicate instanceof ComplexPredicate) {
+    return table.filter(getAttributeFilterFunction(predicate, { booleanAsString: true }));
   }
 
   throw new Error(`Unsupported predicate '${predicate}'`);
+}
+
+/**
+  Checks query on contains restrictions by relationships.
+
+  @method containsRelationships
+  @param {QueryObject} query
+  @return {Boolean}
+*/
+function containsRelationships(query) {
+  let contains = false;
+  if (!query) {
+    return contains;
+  }
+
+  if (query.predicate instanceof SimplePredicate || query.predicate instanceof StringPredicate) {
+    contains = Information.parseAttributePath(query.predicate.attributePath).length > 1;
+  }
+
+  if (query.predicate instanceof DetailPredicate) {
+    contains = true;
+  }
+
+  if (query.predicate instanceof ComplexPredicate) {
+    query.predicate.predicates.forEach((predicate) => {
+      if (predicate instanceof SimplePredicate || predicate instanceof StringPredicate) {
+        contains = Information.parseAttributePath(predicate.attributePath).length > 1;
+      } else if (predicate instanceof DetailPredicate) {
+        contains = true;
+      } else {
+        contains = containsRelationships({ predicate });
+      }
+    });
+  }
+
+  if (query && query.order) {
+    for (let i = 0; i < query.order.length; i++) {
+      let attributePath = query.order.attribute(i).name;
+      if (Information.parseAttributePath(attributePath).length > 1) {
+        contains = true;
+      }
+    }
+  }
+
+  return contains;
 }