+++ addon/services/dexie.js
@@ -1,10 +1,13 @@
-import Ember from 'ember';
-import Dexie from 'npm:dexie';
-import Queue from '../utils/queue';
-
 /**
   @module ember-flexberry-data
 */
+
+import Ember from 'ember';
+import Dexie from 'npm:dexie';
+import Queue from '../utils/queue';
+import isEmbedded from '../utils/is-embedded';
+
+const { isArray, get, merge } = Ember;
 
 /**
   Service for storing [Dexie](https://github.com/dfahlander/Dexie.js) instance for application.
@@ -15,6 +18,14 @@ import Queue from '../utils/queue';
   @public
 */
 export default Ember.Service.extend(Ember.Evented, {
+  /**
+    Contains instances of Dexie.
+
+    @property _dexie
+    @type Object
+  */
+  _dexie: {},
+
   /**
     Count of objects that should be synced down.
 
@@ -66,17 +77,91 @@ export default Ember.Service.extend(Ember.Evented, {
     @return {Dexie} Dexie database.
   */
   dexie(dbName, store, options) {
-    if (!Ember.isNone(this._dexie)) {
-      return this._dexie;
+    let dexie = this.get('_dexie')[dbName];
+    if (dexie instanceof Dexie) {
+      return dexie;
     }
 
-    let db =  new Dexie(dbName, Ember.merge({}, options));
-    let schemas = store.get(`offlineSchema.${dbName}`);
+    let db =  new Dexie(dbName, merge({}, options));
+    let schemas = store.get('offlineSchema')[dbName];
     for (let version in schemas) {
       db.version(version).stores(schemas[version]);
     }
 
-    return this.set('_dexie', db);
+    db.tables.forEach((table) => {
+      let TableClass = table.defineClass({});
+      let modelClass = store.modelFor(table.name);
+      let relationshipNames = get(modelClass, 'relationshipNames');
+      let relationshipsByName = get(modelClass, 'relationshipsByName');
+      let primaryKeyNameFromSerializer = store.serializerFor(table.name).get('primaryKey');
+      let primaryKeyName = primaryKeyNameFromSerializer ? primaryKeyNameFromSerializer : 'id';
+
+      TableClass.prototype.loadRelationships = function(projection) {
+        let promises = [];
+        let relationshipsToIterate = Ember.A();
+
+        if (projection && typeof projection === 'string') {
+          projection = get(modelClass, 'projections').get(projection);
+        }
+
+        if (projection) {
+          // Killing two birds with one stone.
+          for (let attributeName in this) {
+            if (this.hasOwnProperty(attributeName)) {
+              if (attributeName !== primaryKeyName && !projection.attributes.hasOwnProperty(attributeName)) {
+                delete this[attributeName];
+              }
+
+              if (projection.attributes.hasOwnProperty(attributeName) &&
+              (projection.attributes[attributeName].kind === 'belongsTo' || projection.attributes[attributeName].kind === 'hasMany')) {
+                relationshipsToIterate.pushObject(attributeName);
+              }
+            }
+          }
+        } else {
+          relationshipsToIterate.pushObjects(relationshipNames.belongsTo);
+          relationshipsToIterate.pushObjects(relationshipNames.hasMany);
+        }
+
+        relationshipsToIterate.forEach((name) => {
+          let relationship = relationshipsByName.get(name);
+          let saveRelationship = (hash) => {
+            if (!hash) {
+              throw new Error(`Not found relationship with key '${this[name]}' in '${relationship.type}' table.`);
+            }
+
+            if (relationship.options.inverse) {
+              if (!projection || !get(projection, `attributes.${name}.attributes.${relationship.options.inverse}`)) {
+                hash[relationship.options.inverse] = null;
+              }
+            }
+
+            if (relationship.kind === 'hasMany') {
+              for (let i = 0; i < this[name].length; i++) {
+                if (this[name][i] === hash.id) {
+                  this[name][i] = hash;
+                }
+              }
+            } else if (relationship.kind === 'belongsTo') {
+              this[name] = hash;
+            }
+
+            return hash.loadRelationships(projection && projection.attributes[name]);
+          };
+
+          if (this[name] && !relationship.options.async && isEmbedded(store, modelClass, name)) {
+            let ids = isArray(this[name]) ? this[name] : [this[name]];
+            ids.forEach((id) => {
+              promises.push(db.table(relationship.type).get(id, saveRelationship));
+            });
+          }
+        });
+        return Dexie.Promise.all(promises);
+      };
+    });
+
+    this.get(`_dexie`)[dbName] = db;
+    return db;
   },
 
   /**
@@ -118,9 +203,6 @@ export default Ember.Service.extend(Ember.Evented, {
       return operation(db);
     }
   },
-
-  /* Dexie instance */
-  _dexie: null,
 
   /* Queue for requests to Dexie */
   _queue: Queue.create()